Naive merge of Item classes - phase 1
