+++ src/main/java/uk/gov/register/core/FatEntry.java
@@ -1,7 +1,5 @@
 package uk.gov.register.core;
 
-import uk.gov.mint.Item;
-
 public class FatEntry {
     public final Entry entry;
     public final Item item;
