+++ src/test/java/uk/gov/register/functional/DataUploadFunctionalTest.java
@@ -13,11 +13,11 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.RuleChain;
 import org.junit.rules.TestRule;
+import uk.gov.register.core.Item;
 import uk.gov.register.functional.app.CleanDatabaseRule;
 import uk.gov.register.functional.db.TestDBItem;
 import uk.gov.register.functional.db.TestRecord;
 import uk.gov.register.core.Entry;
-import uk.gov.mint.Item;
 import uk.gov.register.RegisterApplication;
 import uk.gov.register.RegisterConfiguration;
 import uk.gov.register.util.CanonicalJsonMapper;
